Clean up openai function calling example
