+++ examples/next-openai/app/api/chat-with-functions/route.ts
@@ -1,5 +1,6 @@
 import { Configuration, OpenAIApi } from 'openai-edge'
 import { OpenAIStream, StreamingTextResponse } from 'ai'
+import { ChatCompletionFunctions } from 'openai-edge/types/api'
 
 // Create an OpenAI API client (that's edge friendly!)
 const config = new Configuration({
@@ -10,8 +11,56 @@ const openai = new OpenAIApi(config)
 // IMPORTANT! Set the runtime to edge
 export const runtime = 'edge'
 
+const functions: ChatCompletionFunctions[] = [
+  {
+    name: 'get_current_weather',
+    description: 'Get the current weather',
+    parameters: {
+      type: 'object',
+      properties: {
+        location: {
+          type: 'string',
+          description: 'The city and state, e.g. San Francisco, CA'
+        },
+        format: {
+          type: 'string',
+          enum: ['celsius', 'fahrenheit'],
+          description:
+            'The temperature unit to use. Infer this from the users location.'
+        }
+      },
+      required: ['location', 'format']
+    }
+  },
+  {
+    name: 'get_current_time',
+    description: 'Get the current time',
+    parameters: {
+      type: 'object',
+      properties: {},
+      required: []
+    }
+  },
+  {
+    name: 'eval_code_in_browser',
+    description: 'Execute javascript code in the browser with eval().',
+    parameters: {
+      type: 'object',
+      properties: {
+        code: {
+          type: 'string',
+          description: `Javascript code that will be directly executed via eval(). Do not use backticks in your response.
+           DO NOT include any newlines in your response, and be sure to provide only valid JSON when providing the arguments object.
+           The output of the eval() will be returned directly by the function.`
+        }
+      },
+      required: ['code']
+    }
+  }
+]
+
 export async function POST(req: Request) {
-  const { messages, functions, function_call } = await req.json()
+  const { messages, function_call } = await req.json()
 
   const response = await openai.createChatCompletion({
     model: 'gpt-3.5-turbo-0613',
